feat: range&nn index integration
1. range query with index
2. nn query with index
