+++ src/datafiles/phase3/nquery1.txt
@@ -1,1 +1,1 @@
-NN(2, target1, 5, N, 1, 2)
+NN(2, target1, 1, Y, 1, 2)
