+++ src/datafiles/phase3/rquery1.txt
@@ -1,1 +1,1 @@
-Range(2, target1, 1, N, 1, 2)
+Range(2, target1, 1, Y, 1, 2)
